% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_utilities.r
\name{.chaste_GHG}
\alias{.chaste_GHG}
\title{Internal function to convert GHG (m-sf) to (cm-sl) and limit value.}
\usage{
.chaste_GHG(GHG)
}
\arguments{
\item{GHG}{Average highest groundwater level, relative to soil surface level, (m) (numeric)}
}
\value{
GHG (cm-sl)
}
\description{
Internal function to convert GHG (m-sf) to (cm-sl) and limit value.
}
