% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_utilities.r
\name{.chaste_GLG}
\alias{.chaste_GLG}
\title{Internal function to convert GLG (m-sf) to (cm-sl) and limit value.}
\usage{
.chaste_GLG(GLG)
}
\arguments{
\item{GLG}{Average lowest groundwater level, relative to soil surface level, (m) (numeric)}
}
\value{
GLG (cm-sl)
}
\description{
Internal function to convert GLG (m-sf) to (cm-sl) and limit value.
}
