% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_utilities.r
\name{.f}
\alias{.f}
\title{Helper function for function '.ht_reduction_brk()'.}
\usage{
.f(x)
}
\arguments{
\item{x}{Numeric vector with four elements:
\describe{
\item{GHG}{Average highest groundwater level, relative to soil surface level (m)}
\item{GLG}{Average lowest groundwater level, relative to soil surface level (m)}
\item{HELP}{(soil) number 1-70}
\item{landuse}{ 1=grassland; 2=arable land}
}}
}
\value{
Numeric vector with reduction in crop production:
\describe{
\item{1}{due to waterlogging (\%)}
\item{2}{due to drought (\%)}
\item{3}{total (\%)}
}
}
\description{
Helper function for function '.ht_reduction_brk()'.
}
