% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_utilities.r
\name{.fdr}
\alias{.fdr}
\title{Function to calculate the reduction in crop production caused by drought.}
\usage{
.fdr(x, GLG)
}
\arguments{
\item{x}{Parameters (A-E) in the analytic function to calculate the reduction in crop production caused by drought.}

\item{GLG}{Average lowest groundwater level, relative to soil surface level, (m) (numeric)}
}
\value{
Reduction in crop production caused by drought (\%); (numeric)
}
\description{
Function to calculate the reduction in crop production caused by drought.
}
