% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_utilities.r
\name{.ftot}
\alias{.ftot}
\title{Function to calculate the total reduction in crop production caused by waterlogging and drought.}
\usage{
.ftot(red_wl, red_dr)
}
\arguments{
\item{red_wl}{Reduction in crop production caused by waterlogging (\%); (numeric)}

\item{red_dr}{Reduction in crop production caused by drought (\%); (numeric)}
}
\value{
Total reduction in crop production caused by both waterlogging and drought (\%); (numeric)
}
\description{
Function to calculate the total reduction in crop production caused by waterlogging and drought.
}
