% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_utilities.r
\name{.help_table_values}
\alias{.help_table_values}
\title{Retreive tabulated values of HELP 1987 table}
\usage{
.help_table_values(HELP, landuse, aard)
}
\arguments{
\item{HELP}{HELP (soil) number 1-70 (integer).}

\item{landuse}{1=grassland; 2=arable land (integer)}

\item{aard}{Nature of the reduction in crop production ("Natschade", "Droogteschade") (character)}
}
\value{
help_table_values (numeric vector)
}
\description{
Retreive tabulated values of HELP 1987 table
}
