% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_utilities.r
\name{.rmse_dr}
\alias{.rmse_dr}
\title{Calulate the root-mean-square error in the relation between the tabulated HELP1987 values and calculated values
of reduction in crop production caused drought.}
\usage{
.rmse_dr(x, HELP, landuse)
}
\arguments{
\item{x}{Parameters (A-E) in the analytic function to calculate the reduction in crop production caused by drought.}

\item{HELP}{HELP (soil) number 1-70 (integer).}

\item{landuse}{1=grassland; 2=arable land (integer)}
}
\value{
Root-mean-square error (numeric)
}
\description{
The residual standard error is the positive square root of the mean square error.
}
