% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_utilities.r
\name{.rmse_wl}
\alias{.rmse_wl}
\title{Calulate the root-mean-square error in the relation between the tabulated HELP1987 values and the calculated values
of the reduction in crop production caused by waterlogging.}
\usage{
.rmse_wl(x, HELP, landuse)
}
\arguments{
\item{x}{Parameters (A-E) in the analytic function to calculate the reduction in crop production caused by waterlogging.}

\item{HELP}{HELP (soil) number 1-70 (integer).}

\item{landuse}{1=grassland; 2=arable land (integer)}
}
\value{
Root-mean-square error (numeric)
}
\description{
Calulate the root-mean-square error in the relation between the tabulated HELP1987 values and the calculated values
of the reduction in crop production caused by waterlogging.
}
