% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HELP_map_NL2020}
\alias{HELP_map_NL2020}
\title{Raster map of 1987 HELP codes for the Netherlands}
\format{
An object of class \code{RasterLayer} of dimension 12494 x 10580 x 1.
}
\source{
\href{https://www.wur.nl/nl/show/bodemfysische-eenhedenkaart-bofek2020.htm}{Bofek2020}
}
\usage{
HELP_map_NL2020
}
\description{
Raster map of 1987 HELP codes for the Netherlands, based on the Bofek2020 raster map.
Resolution 25x25m.
}
\details{
The HELP numbers 71 and 72 are converted to the most similar soil codes in the HELP-table (1987).
\href{https://library.wur.nl/WebQuery/wurpubs/fulltext/27040}{HELP-2005 Stowa rapport 2005/16 (page 16)}
HELP number 71 --> 67; HELP number 72 --> 60.
}
\keyword{datasets}
