% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HELP_table_NL2020}
\alias{HELP_table_NL2020}
\title{Raster of HELP codes for the Netherlands}
\format{
An object of class \code{RasterLayer} of dimension 12494 x 10580 x 1.
}
\usage{
HELP_table_NL2020
}
\description{
This dataset contains a raster of HELP codes for the Netherlands.
Resolution 25x25m
}
\keyword{datasets}
