% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlptabel.r
\name{hlptabel}
\alias{hlptabel}
\title{hlptabel: Calculate reduction in crop production caused by waterlogging and drought.}
\description{
Based on the publication "De invloed van de waterhuishouding op de landbouwkundige produktie".
Rapport van de werkgroep HELP-tabel, Mededelingen Landinrichtingsdienst 176 (1987).
\href{https://edepot.wur.nl/188152}{Rapport van de werkgroep HELP-tabel}
}
\details{
Approximate analytical formulas are used to calculate the reduction in crop productions.

The following functions are exported:
\itemize{
\item \code{\link{ht_reduction}}: Calculate reduction in crop production caused by waterlogging and drought.
\item \code{\link{ht_reduction_brk}}: Calculate reduction in crop production caused by waterlogging and drought using a SpatRaster object as input.
\item \code{\link{ht_tab_calc_values}}: Tabulated and calculated reductions in crop production.
\item \code{\link{ht_plot_tab_calc_values}}: Plot of tabulated and calculated reductions in crop production.
}

The following datasets are included in the package:
\itemize{
\item \strong{HELP_map_NL2020}: HELP map of the Netherlands based on the Bofek2020 map.
\item \strong{landuse_map_NL2021}: Landuse map of the Netherlands based on the top25raster map.
}

To use these datasets, the following code can be used:
\itemize{
\item \code{HELP_map_NL2020 <- file.path( find.package("hlptabel"), "extdata", "HELP_map_NL2020.tif") |> terra::rast()}
\item \code{landuse_map_NL2021 <- file.path( find.package("hlptabel"), "extdata", "landuse_map_NL2021.tif") |> terra::rast()}
}

You might consider to reclassify the HELP numbers 71 and 72 to the most similar soil codes in the HELP-table (1987).
\href{https://library.wur.nl/WebQuery/wurpubs/fulltext/27040}{HELP-2005 Stowa rapport 2005/16 (page 16)}
HELP number 71 --> 67; HELP number 72 --> 60.
Dataset \strong{HELP_map_NL2020} is already converted in this way.
}
