% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_utilities.r
\name{ht_bofek_numbers}
\alias{ht_bofek_numbers}
\title{Valid bofek numbers.}
\usage{
ht_bofek_numbers(version = 2012)
}
\arguments{
\item{version}{Bofek version (2012 or 2020).}
}
\value{
bofek (integer)
ht_bofek_numbers()
}
\description{
Valid bofek numbers.
}
