% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_utilities.r
\name{ht_bofek_to_HELPnr}
\alias{ht_bofek_to_HELPnr}
\title{Get HELP number by specifying a bofek number.}
\usage{
ht_bofek_to_HELPnr(bofek)
}
\arguments{
\item{bofek}{Bofek number(integer).}
}
\value{
HELP HELP (soil) number 1-70 (integer).
}
\description{
Get HELP number by specifying a bofek number.
}
\examples{
ht_bofek_to_HELPnr( 101 )
}
