% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_utilities.r
\name{ht_HELPnr_to_HELPcode}
\alias{ht_HELPnr_to_HELPcode}
\title{Get HELP (soil) code by specifying HELP number.}
\usage{
ht_HELPnr_to_HELPcode(HELP)
}
\arguments{
\item{HELP}{HELP (soil) number 1-70 (integer).}
}
\value{
HELPcode (character).
}
\description{
Get HELP (soil) code by specifying HELP number.
}
\examples{
ht_HELPnr_to_HELPcode( HELP=15 )
}
