% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_utilities.r
\name{ht_plot_tab_calc_values}
\alias{ht_plot_tab_calc_values}
\title{Return plot of tabulated and calculated reductions in crop production.}
\usage{
ht_plot_tab_calc_values(HELP, landuse)
}
\arguments{
\item{HELP}{HELP (soil) number 1-70 (integer).}

\item{landuse}{1=grassland; 2=arable land (numeric)}
}
\value{
plot with tabulated and calculated reductions in crop production.
}
\description{
Return plot of tabulated and calculated reductions in crop production.
}
\examples{
ht_plot_tab_calc_values(HELP=15, landuse=1)
ht_plot_tab_calc_values(HELP=15, landuse=1)
}
