% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_utilities.r
\name{ht_reduction_brk}
\alias{ht_reduction_brk}
\title{Calculate reduction in crop production caused by waterlogging and drought.}
\usage{
ht_reduction_brk(x)
}
\arguments{
\item{x}{A Spatraster object with the following layers (in this order):
\describe{
\item{GHG}{Average highest groundwater level, relative to soil surface level (m)}
\item{GLG}{Average lowest groundwater level, relative to soil surface level (m)}
\item{HELP}{(soil) number 1-70}
\item{landuse}{ 1=grassland; 2=arable land}
}}
}
\value{
A Spatraster object with the following layers:
\describe{
\item{wl}{Reduction in crop production caused by waterlogging (\%)}
\item{dr}{Reduction in crop production caused by drought (\%)}
\item{tot}{Total reduction in crop production caused by both waterlogging and drought (\%)}
\item{landuse}{ 1=grassland; 2=arable land}
}
}
\description{
Calculate reduction in crop production caused by waterlogging and drought.
}
\details{
Results are created using the function ht_reduction()'.
}
\examples{
x <- terra::rast(system.file("extdata","example_spatraster.tif",package="hlptabel"))
r <- ht_reduction_brk(x)
}
