% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_utilities.r
\name{ht_soil_unit_to_HELPnr}
\alias{ht_soil_unit_to_HELPnr}
\title{Get HELP number by specifying a soil unit ("soil_unit").}
\usage{
ht_soil_unit_to_HELPnr(soil_unit)
}
\arguments{
\item{soil_unit}{Soil unit (character).}
}
\value{
HELP HELP (soil) number 1-70 (integer).
ht_soil_unit_to_HELPnr( "faVzt" )
}
\description{
Get HELP number by specifying a soil unit ("soil_unit").
}
