% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_utilities.r
\name{ht_soilnr_to_HELPnr}
\alias{ht_soilnr_to_HELPnr}
\title{Get HELP number by specifying a soil number (1010, ..., 22020).}
\usage{
ht_soilnr_to_HELPnr(soilnr)
}
\arguments{
\item{soilnr}{Soil number (1010, ..., 22020) (integer).}
}
\value{
HELP HELP (soil) number 1-70 (integer).
ht_soilnr_to_HELPnr( 1030 )
}
\description{
Get HELP number by specifying a soil number (1010, ..., 22020).
}
