% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_utilities.r
\name{ht_tab_calc_values}
\alias{ht_tab_calc_values}
\title{Return tabulated and calculated reductions in crop production.}
\usage{
ht_tab_calc_values(HELP, landuse, aard)
}
\arguments{
\item{HELP}{HELP (soil) number 1-70 (integer).}

\item{landuse}{1=grassland; 2=arable land (integer)}

\item{aard}{Nature of the reduction in crop production ("Natschade", "Droogteschade") (character)}
}
\value{
Data frame with the fields:
HELP: HELP (soil) number 1-70 (integer).
landuse: 1=grassland; 2=arable land (numeric)
aard: Nature of the reduction in crop production ("Natschade", "Droogteschade") (character)
GHG: Average highest groundwater level, relative to soil surface level, (m) (numeric)
GLG: Average lowest groundwater level, relative to soil surface level, (m) (numeric)
red_HELP_table: Reduction in crop production in HELP 1987 table (\%); (numeric)
red_calculated: Calculated reduction in crop production (\%); (numeric)
}
\description{
Return tabulated and calculated reductions in crop production.
}
\examples{
ht_tab_calc_values(HELP=15, landuse=1, aard="Natschade")
ht_tab_calc_values(HELP=15, landuse=1, aard="Droogteschade")
}
