% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landuse_map_NL2021}
\alias{landuse_map_NL2021}
\title{landuse map (units gras=1, akker=2) for the Netherlands.}
\format{
An object of class \code{RasterLayer} of dimension 13000 x 10800 x 1.
}
\usage{
landuse_map_NL2021
}
\description{
Raster map (resolution 25x25m) of landuse (2021) based on the TOP25raster_GEOTIFF dataset 2021.
}
\keyword{datasets}
