% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HELP_map_NL2020}
\alias{HELP_map_NL2020}
\title{Raster map of 1987 HELP codes for the Netherlands}
\format{
An object of class \code{RasterLayer} of dimension 12494 x 10580 x 1.
}
\usage{
HELP_map_NL2020
}
\description{
Raster map of 1987 HELP codes for the Netherlands, based on the Bofek2020 raster map.
Resolution 25x25m.
}
\keyword{datasets}
