% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlptabel.r
\docType{package}
\name{hlptabel}
\alias{hlptabel}
\title{hlptabel: Calculate reduction in crop production caused by waterlogging and drought.}
\description{
Based on the publication "De invloed van de waterhuishouding op de landbouwkundige produktie".
Rapport van de werkgroep HELP-tabel, Mededelingen Landinrichtingsdienst 176 (1987).
\href{https://edepot.wur.nl/188152}{Rapport van de werkgroep HELP-tabel}
}
\details{
Approximate analytical formulas are used to calculate the reduction in crop productions.

https://r-pkgs.org/data.html

This package exports the following functions:
\itemize{
\item \code{\link{ht_reduction}}: Calculate reduction in crop production caused by waterlogging and drought.
\item \code{\link{ht_reduction_brk}}: Calculate reduction in crop production caused by waterlogging and drought using a RasterBrick object as input.
\item \code{\link{ht_tab_calc_values}}: Tabulated and calculated reductions in crop production.
\item \code{\link{ht_plot_tab_calc_values}}: Plot of tabulated and calculated reductions in crop production.
}

You might consider to reclassify the HELP numbers 71 and 72 to the most similar soil codes in the HELP-table (1987).
HELP number 71 --> 67; HELP number 72 --> 60. Dataset HELP_map_NL2020 is already converted in this way.
}
