% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlptabel.r
\docType{package}
\name{hlptabel}
\alias{hlptabel}
\title{hlptabel: Calculate reduction in crop production caused by waterlogging and drought.}
\description{
Based on the publication "De invloed van de waterhuishouding op de landbouwkundige produktie".
Rapport van de werkgroep HELP-tabel, Mededelingen Landinrichtingsdienst 176 (1987).
\href{https://edepot.wur.nl/188152}{Rapport van de werkgroep HELP-tabel}
}
\details{
Approximate analytical formulas are used to calculate the reduction in crop productions.

https://r-pkgs.org/data.html

This package exports the following functions:
\itemize{
\item \code{\link{ht_reduction}}: Calculate reduction in crop production caused by waterlogging and drought.
\item \code{\link{ht_reduction_brk}}: Calculate reduction in crop production caused by waterlogging and drought using a RasterBrick object as input.
\item \code{\link{ht_soilnr_to_HELPnr}}: Get HELP number by specifying a soil number (1010, ..., 22020).
\item \code{\link{ht_bofek_to_HELPnr}}: Get HELP number by specifying a bofek number.
\item \code{\link{ht_soil_unit_to_HELPnr}}: Get HELP number by specifying a soil unit ("soil_unit").
\item \code{\link{ht_soil_units}}: Valid soil units.
\item \code{\link{ht_bofek_numbers}}: Valid bofek numbers.
\item \code{\link{ht_HELPnr_to_HELPcode}}: Get HELP (soil) code by specifying HELP number.
\item \code{\link{ht_tab_calc_values}}: Tabulated and calculated reductions in crop production.
\item \code{\link{ht_plot_tab_calc_values}}: Plot of tabulated and calculated reductions in crop production.
}

You might consider to reclassify the HELP numbers 71 and 72 to the most similar soil codes in the HELP-table (1987).
if \code{x} is the raster with HELP numbers (1-72), reclassify with:

\code{x[x[]==71] <- 67}

\code{x[x[]==72] <- 60}
}
