% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_utilities.r
\name{ht_reduction_brk}
\alias{ht_reduction_brk}
\title{Calculate reduction in crop production caused by waterlogging and drought using a RasterBrick object as input.}
\usage{
ht_reduction_brk(x)
}
\arguments{
\item{x}{A RasterBrick object with the following Raster layers (in this order):
\describe{
\item{GHG}{Average highest groundwater level, relative to soil surface level (m)}
\item{GLG}{Average lowest groundwater level, relative to soil surface level (m)}
\item{HELP}{(soil) number 1-70}
\item{landuse}{ 1=grassland; 2=arable land}
}}
}
\value{
A RasterBrick object with the following Raster layers:
\describe{
\item{wl}{Reduction in crop production caused by waterlogging (\%)}
\item{dr}{Reduction in crop production caused by drought (\%)}
\item{tot}{Total reduction in crop production caused by both waterlogging and drought (\%)}
\item{landuse}{ 1=grassland; 2=arable land}
}
}
\description{
Calculate reduction in crop production caused by waterlogging and drought using a RasterBrick object as input.
}
\examples{
x <- raster::brick(system.file("extdata","example_brick.grd",package="hlptabel"))
raster::beginCluster(exclude="hlptabel")
r <- raster::clusterR(x,ht_reduction_brk)
raster::endCluster()
}
